added float casting

+++ roundtrip/selectTopK.py
@@ -9,7 +9,7 @@ if (len(sys.argv) == 1):
 if len(sys.argv) > 1:
     inputFile = sys.argv[1]
 if len(sys.argv) > 2:
-    percentage = str((double(sys.argv[2] + ".0") / 100.0))
+    percentage = str((float(sys.argv[2] + ".0") / 100.0))
 outputFile = inputFile.rsplit('.',1)[0] + "_" + percentage + ".nt"
